+++ me_software/throttle_control.h
@@ -13,7 +13,7 @@ extern const int THROTTLE_SIGNAL_PWM;
 
 extern const int MANUAL_SWITCH_PIN;
 
-extern const int REFERENCE_VOLTAGE;
+extern const double REFERENCE_VOLTAGE;
 extern const int BATTERY_VOLTAGE;
 
 void write_speed(mraa_pwm_context throttle_out, float signal_out);
